Deployment/requirements.txt
Flask==2.3.2
pandas==1.5.3
numpy==1.24.3
google-cloud-bigquery==3.11.0
google-auth==2.22.0
